Add files via upload
This code is © Battaglia et al., 2018, and it is made available under the GPL license enclosed with the software.

Over and above the legal restrictions imposed by this license, if you use this software for an academic publication then you are obliged to provide proper attribution. This can be to this code directly.

Battaglia S., Garofalo S., di Pellegrino G., Context-dependent extinction of threat memories: influences of healthy aging. Scientific Reports, 2018, (8) 12592.